[MAINT] rename typing module
