+++ nilearn/nilearn_typing.py
