+++ nilearn/plotting/image/img_plotting.py
@@ -45,6 +45,16 @@ from nilearn.image import (
 )
 from nilearn.maskers import NiftiMasker
 from nilearn.masking import apply_mask, compute_epi_mask
+from nilearn.nilearn_typing import (
+    Annotate,
+    ColorBar,
+    DisplayMode,
+    DrawCross,
+    OutputFile,
+    Radiological,
+    ResamplingInterpolation,
+    Title,
+)
 from nilearn.plotting import cm
 from nilearn.plotting._engine_utils import create_colormap_from_lut
 from nilearn.plotting._utils import (
@@ -56,16 +66,6 @@ from nilearn.plotting.displays import OrthoSlicer, get_projector, get_slicer
 from nilearn.plotting.displays._slicers import save_figure_if_needed
 from nilearn.plotting.image.utils import MNI152TEMPLATE, load_anat
 from nilearn.signal import clean
-from nilearn.typing import (
-    Annotate,
-    ColorBar,
-    DisplayMode,
-    DrawCross,
-    OutputFile,
-    Radiological,
-    ResamplingInterpolation,
-    Title,
-)
 
 
 def show() -> None:
